src/python/automation/requirements.txt
GitPython==3.1.11
gql==2.0.0
py4j==0.10.9.1
python-dateutil==2.8.1
pytz
git+https://github.com/GNS-Science/nshm-toshi-client#egg=nshm_toshi_client
